♻️ refactor: improve code readability
