+++ src/features/PostLengthFilter/lib/filterByLength.ts
@@ -1,12 +1,6 @@
 import type { Post } from "@/entities/post/model/types";
 
-type FilterByLengthParams = {
-  posts: Post[];
-  min?: number;
-  max?: number;
-};
-
-export const filterByLength = ({ posts, min, max }: FilterByLengthParams) =>
+export const filterByLength = (posts: Post[], min?: number, max?: number) =>
   posts.filter((post) => {
     const length = post.title.length;
     const meetsMin = min === undefined || length >= min;
